Restructure how content is linked to users
De-normalize the database to store the last content that users received
on the user. This makes it easier to know what to send the user next,
and reduces the reliance on their age in months. This was causing issues
where the child would turn a month older at the beginning of each month,
and depending on when they had signed up it meant they were missing
content.